backend/medical-backend/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
python-dotenv==1.0.1
google-generativeai>=0.8.7
scikit-learn==1.5.0
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
faker==25.0.0
pydantic==2.7.1
